+++ mosaic-runtime/src/commonMain/kotlin/com/jakewharton/mosaic/ui/Static.kt
@@ -3,10 +3,6 @@
 package com.jakewharton.mosaic.ui
 
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableIntStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.runtime.snapshots.SnapshotStateList
 import com.jakewharton.mosaic.modifier.Modifier
 import kotlin.jvm.JvmName
@@ -20,8 +16,10 @@ public fun <T> Static(
 	items: SnapshotStateList<T>,
 	content: @Composable (T) -> Unit,
 ) {
-	var lastDrawn by remember { mutableIntStateOf(0) }
-	var lastRendered by remember { mutableIntStateOf(0) }
+	// These are not state because we don't want updates to them to trigger recomposition. Changes to
+	// the list are the only thing which should trigger recomposition.
+	var lastDrawn = 0
+	var lastRendered = 0
 
 	Node(
 		measurePolicy = { measurables, constraints ->